Move encoding helper methods into RSpec::Support::EncodingHelpers
